Extract a Committer and RunloopRebalanceListener from the Runloop
